Fix login bug in unit tests
This was a silly bug. It's common knowledge that you should never directly
set the `password` field on a Django user object; instead, you should use
either `create_user` or use `set_password`, because the password is stored
hashed. Unfortunately, that for loop iterated over all the keyword arguments,
INCLUDING password, and set them on the user.

This adds a set of kwargs to skip (since they've been handled already), and
the unit test passes.