Add a 'dependencies' task to the fabfile.
This adds a 'dependencies' task that checks the contents of your
requirements.txt file against the output of `pip freeze`. If there
are mismatches, it outputs a table with the required and currently
installed versions of the modules.

This task will automatically run after creating a new feature
branch (fab feature) and after rebasing onto the latest upstream
(fab update).