Add dummy Login view.
This adds a basic login view at /www that looks identical to the
existing login view at /login. It sends the login request to the
API and correctly handles errors, but doesn't do anything on
success except print the response data to the console.

The next step would be to add a CurrentUserStore to... store the
current user. We then have a model of whether the user is logged
in.