Add Flux components for the current user.
This adds bare-bones versions of the fundamental components in
Flux: a Store, the Dispatcher, and Actions. This completes the
loop in Flux's one-directional data flow: When a successful login
response comes back to the login view,
- the view sends a login action, which sends it to the dispatcher.
- The store subscribes to the dispatcher for the action, updates
internal state, and emits a change event.
- The application listens for the change event, and updates accordingly.
(in this case, it just logs to the console).