Add a dummy AdvisorProfileView.
This adds an AdvisorProfile view that does nothing but provide a
button to log out. This completes the basic frame of logging in
and logging out.

Still to do (other than actually building the profile view) is to
enforce that the user is logged-in and an advisor. Ideally, we want
to do this as high-up in the component hierarchy as possible, so
any sub-components can just receive the current user as a prop and
not have to worry about it.