(#174) Bootstrap current user data on page load.
This adds a `currentUser` variable to the template that's assigned
the serialized user data. Then, the first thing we do in the main
script is to bootstrap the CurrentUserStore. We also assert that
the store is bootstrapped before allowing any methods on it, but
I'm not sure if that's necessary.

In order for this to work, we have to use the `safe` filter in the
template. This could be dangerous, since the object contains user-
submitted data, so it's important that we thoroughly validate when
users and schools are created.