(#175) Add RedirectView component for redirecting on page load.
This introduces the RedirectView, which is displayed for the root
path. If the user is logged-in, it will redirect to the advisor
profile view. Otherwise, it will redirect to the login view.

This also removes react-router-component in favor of RRouter, its
successor. This required a bit of a change to the code, as all the
routes are now done in the main module instead of in the Huxley
app module.