Clean up dead code in huxley.accounts.
This removes a bunch of the templates, template tags, and views
in the accounts app. It leaves login-as-user and logout, just
in case. It also scraps all the static files, but moves the
SCSS to www (because it's still used).

It removes the useless `login` method from the model, and renames
the test to `test_user.py`.

Test Plan: Login and logout of the app, verify it works.