Add phone number auto-formatting to RegistrationView.
This adds a utility function to automatically format phone numbers
for the primary/secondary phone fields in RegistrationView. It takes
a string as input (possibly partially formatted), and outputs a string
that adheres to the format (DDD) DDD-DDDD xDDDDD.

Test Plan: Added a unit test, inputted some phone numbers in RegistrationView.

Resolves #223, Closes #271.