Fix production DB pollution bug in huxley.core.tests.
Django tests are run with dummy database, so that production data
isn't touched. However, it's only set up after modules are loaded.
So... if you execute code that touches the DB before the test DB
is created, you'd hit production data.

News for me: apparently, expressions assigned to class members are
executed when the module is loaded! So, I realized I had literally
hundreds of countries named Lolville, etc.

This sets things up in the `setUp` function, like it's supposed to.