Update to Django 1.8
This updates to 1.8 so we can iron out any bugs before moving onto 1.9. I discovered that the Django User model now validates uniqueness and character set, so we no longer have to do that in the serializer.

Other than that, just needed to fix minor type issues, like enforcing that School's fees_paid and fees_owed fields were consistently using Decimal instances rather than ints or floats. Perhaps we should reconsider and use just an int field instead, but that's a discussion for a later time!

Resolves #443.