Fix Spanish speaking & Chinese speaking delegate issue (#483)
This modifies the SchoolSerializer's validate method to collect all validation issues into a single ValidationError with a dict of every field checked by the method with an issue. It introduces a check on whether the # of Spanish or Chinese speaking delegates exceeds the total # of delegates, as well as fix a bug where an invalid phone number, e.g. (123) 456 78, would raise a ValidationError but the error would not appear in registration view.

Fixes #483.