Introduce python autoformatter (#496)
This adds a python auto formatter as discussed in #472. I introduced a command 'fab format' to run the auto formatter, which also gets run as part of submit. As an example of what the auto formatter might do, look at how the code of init.py and git.py was modified.

Resolves #472.