Add pylint to workflow (#528)
This adds Pylint to the workflow. Upon submit users will have the option to view a brief Pylint evaluation of the python files that were changed. They can then choose to continue with the submit based off the evaluation. This only prints an evluation, it does not alter the code in anyway (which is why I seperated it from the format command).

Fixes #294.