Fix text refocus (#530)
This is the simplest resolution to the issue with text refocus discussed in #519. By leaving text inputs as uncontrolled components, the user's input is still preserved internally but the JS no longer resets the value every render. What was happening was the user would type, React would handle the change & update the state, then it would re-render and in doing so look at the value prop for that input (which was set to retrieve its data from the state), obliterate what the user had typed and replace it with the state (which was the same as what the user had typed). This meant cursor position was not being preserved. This just removes the explicit setting of the value prop (which doesn't change how the state is updated, just how the user sees it).

Fixes #519.