Merge SchoolAssignments into AssignmentList (#547, #548)
This should not be a separate API view; it should instead be a filter on AssignmentList. This merges the two and deletes SchoolAssignments.

Permissions will need to be a bit more complex, so instead of completely general-purpose Permission classes, this proposes a Permission class per-view (when needed), which then delegates to common functions for code reuse.

Tests are updated to verify the filtering behavior, and I inspected the request/response manually in the JS console in the browser.