Set correct contentType on AJAX requests
This fixes the need to do the QueryDict hack because it correctly sets the content type on the request, which causes it to be parsed as JSON instead of url-encoded data.

This also adds the JSON stringification in ServerAPI, so we don't have to do it manually in the calls.