Only save summaries that have been modified (#579)
This makes it so that when a chair saves summaries, only the delegates that have had their summaries updated get sent to the backend. This helps with letting multiple chairs access the page at once, as now only changes made to the same delegates' summaries would get clobbered (versus before, where every summary for the committee would get clobbered).