Initial setup for LESS
This sets up the infrastructure for using LESS to compile our CSS, with the eventual goal of getting off of Django Pipeline. Changes in this PR:
- Sets up all the loaders in Webpack
- Converts one file as an example
- Changes the CSS style to more closely match JS (two-space indentation, no newline for first brace)

Notice that our CSS files are now colocated with their corresponding React components, instead of in their own directory. This seems fundamentally easier to understand.

Eventually, once we're all converted to LESS, we'll be able to take advantage of its features if we choose, like mixins and variables.